Added on delete cascade en las tablas Nomina y NominaResumen
Esto resuelve el requisito de eliminar los registros de las tablas NominaResumen y NominaDetalle cuando se elimina una Nómina.
También se añadió desde la API un condicional que evita que se eliminen las Nóminas contabilizadas.

+++ NominaAPI/NominaAPI/Controllers/NominaController.cs
@@ -204,6 +204,10 @@ namespace NominaAPI.Controllers
             {
                 return NotFound();
             }
+            if (nomina.Contabilizado)
+            {
+                return BadRequest("Esta nómina ya está contabilizada, no se puede eliminar");
+            }
 
             db.Nomina.Remove(nomina);
             await db.SaveChangesAsync();
